ncompress: update download sha
This was failing to build from source soley because the sha256 was
incorrect for the tarball.

I don't have any direct knowledge of thie history of this package,
but I suspect what happened is that the git history got rewritten
per the comment here: https://github.com/vapier/ncompress/issues/26#issuecomment-570867943
and that meant that github's generated tarball got changed slightly.