cpmtools: ressurect with 2.21
This was just marked deprecated, but it actually builds/runs OK

There was a previous attempt a few months ago by @alebcay to upgrade
this pacakge to 2.21 but it failed due to "brew test" failing.  (#61617)

However, I think the failure was not due to the upgrade but by
@fxcoudert's remval of its "libdsk" option (#32207) which made what
had been an optional component required.  However, when built in this
mode you need to pass an extra "-T raw" to several utilities or they
don't always work.