dupseek: fix test and undeprecated
There was nothing wrong with this tool on Big Sur.  The only issue
is that the test assumed incorrectly that readdir() would list files
in creation-order.  This is not true on APFS at least.  Therefore
when testing if two files are duplicates, either one could be detected
as the "dup"