pike: workarounds for building on Xcode 12
Issue also reported upstream.

Also needed to add a dependency on "librsvg" to get it to build
correctly.  This is probably just because I already had it installed
so at it detected its availability but then not all of the bits
made it into the sandbox.  However, there are a bunch of sub-modules
that don't build with our pike install that probably *could* if we
added more dependencies.  Someone with more familiarity with this
package might want to add more such dependencies.