gnu-prolog: switch to URL with original package
This formula had been using the download pointed to by the project's
homepage, with a fallback URL on debian.org.  Both of those were
different than the sha256 expected, and were different from each other
as well.

However, the version on gnu.org did match the sha256.  So what gives?

Looking at the tree tarballs it's clear that the one from debian.org
isn't even close to the right size so we can rule that out immediately:

.rw-r--r-- 3.6M mitch  2 Oct  8:23 homepage/gprolog-1.4.5.tar.gz
.rw-r--r-- 1.1M mitch 11 Mar  2019 deb/gprolog_1.4.5.0.orig.tar.gz
.rw-r--r-- 3.6M mitch  7 Feb  2019 gnu/gprolog-1.4.5.tar.gz

Comparing the GNU tarball to the "homepage" tarball, it looks like
recently the author pushed a couple minor changes to the project
(and regenerated html documentation) while still calling it 1.4.5.
Notably:

--- gnu/gprolog-1.4.5/src/EnginePl/gprolog_cst.h        2018-07-13 17:28:27.000000000 +0100
+++ homepage/gprolog-1.4.5/src/EnginePl/gprolog_cst.h   2020-10-02 08:18:41.000000000 +0100
@@ -8,8 +8,8 @@
 #define PROLOG_DIALECT         "gprolog"
 #define PROLOG_NAME            "GNU Prolog"
 #define PROLOG_VERSION         "1.4.5"
-#define PROLOG_DATE            "Jul 13 2018"
-#define PROLOG_COPYRIGHT       "Copyright (C) 1999-2018 Daniel Diaz"
+#define PROLOG_DATE            "Oct  2 2020"
+#define PROLOG_COPYRIGHT       "Copyright (C) 1999-2020 Daniel Diaz"

[...]

--- gnu/gprolog-1.4.5/ChangeLog 2018-07-14 09:19:12.000000000 +0100
+++ homepage/gprolog-1.4.5/ChangeLog    2018-10-23 16:36:00.000000000 +0100
@@ -1,3 +1,7 @@
+Tue Oct 23 17:34:21 CEST 2018 <Daniel.Diaz@univ-paris1.fr>
+
+       * fix problem with old gcc (gcc < 6 does not produce PIE code by default)
+

I didn't see anything nefarious inserted.  Anyway, for now just switch
the download URL to gnu.org so we can keep building from the source
we were happy with before.