xdotool: avoid problems with CLOCK_MONOTONIC redefinition
This patch was originally proposed by @claui in 2017 (see #19639)
but there were concerns with the fix not working on pre-10.12
versions of OS/X.  Instead another fix regarding $SDKROOT was
made.

This fix doesn't seem to work anymore, but I don't think the
issue with the original fix is an issue any more (all Homebrew-supported
versions of OS/X should have a real CLOCK_MONOTONIC in their SDK)
so going back to that fix.

Note that there is a bug in the previous fix where the comparison
should have been "Xcode.version.to_i >= 8" but just unsetting
$SDKROOT doesn't seem enough to fix the compilation issue anyway
so that's irrelevant.