lemon: remove dead download URL
Just promote the mirror to be the main URL since I can't find
a better source for this tarball.

It's not clear what the long-term future for this formula is since
looking at the project's webpage it seems like they aren't
in the business of distributing tarballs but instead just point
interested parties at files shipped as part of SQLite.

One possibility is to just switch this formula to downloading
the SQLite tarball and building lemon.c from there, but that still
leaves the question of what version number to use?