orbit: deparallelize build but also deprecate
The IDL compiler was dying with a Segmentation Fault, but it seemed
to be happening at the same time some of its input files were being
written to.  Compiles fine for me once I disable parallel building.

However, further research shows that this is very deprecated.. in
fact the deprecation process started in 2007 on the GNOME side meaning
that when this formula was added in 2012 it was already well past
its sell-by date.  Now even the upstream web page is gone, so it's
probably safe to start the long goodbye from Homebrew.