snappystream: build requires C++11
This must have worked at one point, and I think upstream is actually
OK.  However *std_cmake_args overriding of CMAKE_CXX_FLAGS_RELEASE
seems to have changed the default C++ dialect that this is building
with.

Just tell cmake to default to C++11 for this project and it builds
fine again.