whohas: disable formula
I made the changes needed to get "brew test" working again,
but per the discussion on the PR the upstream package is
basically dead at this point.

Since it relies on hard-coded URLs to find info about
distros, some of them are broken and the others are very
out of date.  Probably nobody is getting much valuable
use out of it at this point so let's go straight to disable!