enigma: fix compile on Big Sur
This is another failure caused by MacOS 11's weird dynamic library
system where the /usr/lib/*.dylib aren't physically present on
the filesystem but do exist as dymamic libraries which you can
link against, dlopen, etc.

The configure.ac for this package includes some hacks that put
/usr/lib/libiconv.dylib directly onto some command lines and
that later causes later test programs to fail.  In this case
the end result was that it misdetected the versions of SDL leading
to them being rejected.

This formula already does a bunch of inreplace surgery on
the configure script (for better or worse) so the easiest fix
I think is to just add one more.  With this the game compiles and
runs for me on Big Sur.