kpcli: fix build on Big Sur
This is another example of perl modules being confused by the
lack of /usr/lib/lib*.dylib files being physically present on the
filesystem.  Fix is the same as for notmuch-mutt (#67964)

I also refreshed the other perl modules while here.  The Clipboard
module switched from MakeMaker to Module::Build in the meantime.