libpano: hacks to build on Big Sur
As described in the comment, there is a long-standing issue with
libtiff and Security.framework defining types of the same name.  This
has been reported in multiple forms dating back to at least
  http://bugzilla.maptools.org/show_bug.cgi?id=2464
  https://github.com/Homebrew/homebrew-core/issues/36035
And most recently I started a discussion as part of a related closed bug:
  https://gitlab.com/libtiff/libtiff/-/merge_requests/51#note_476410984

This is now biting libpano on Big Sur, apparently because some system
header is newly sucking in that Security.framework header.  As a
hopefully-temporary hack we can use the preprocessor to change that
symbol name (but only while processing libtiff headers)

The long-term fix will probably be for libtiff not avoid these common
names, but that discussion is ongoing.  This fixes a formula build
in the short/medium term.