extract_url: fix TUI on Big Sur
When called *without* the "-l" option, extract_url is supposed to
be a curses-based app that lets you interactively pick which URL
to open.  However if there is an issue with the perl Curses library
it just silently falls back to "-l" mode.

This is what was happening on Big Sur; since Curses was failing
in the same weird way as in the asciiquarium Forumla (see #66485)
the package appeared to build OK but the curses-based UI didn't actually
work.

I'm importing the same nasty workaround I made to get asciiquarium
happy.  I also added some tests that would catch if this behavior
happened again -- both by testing that our perl resources actually
load OK and by running the TUI in a pty.  The pty test isn't CI-friendly
(at least on the post-10.14 CI environments) but the resource-loading
test should give CI some decent coverage for this condition.

Also bump up the versions of the resources while I'm here, and therefore
bumping the revision.