quex 0.71.0
This was released awhile ago but "brew livecheck" didn't notice it
because it's now only available as a .zip not a .tar.gz

This includes the long-awaited conversion to Python 3.X, but it's
not clear what exact versions are supported.  "brew test" failed
with Python 3.7 but only a minor tweak (which is already in the
upstream HEAD) allowed it to work.  However, I am not 100% sure
that it'll be completely happy.  There is an existing upstream
bug tracking 3.8 support so hopefully this will be cleared up soon.