kite: fix build against bdw-gc >= 8.0
The patch that @fxthomas made in 2012 to allow building against
bdw-gc 7.2 had a bug which broke when bdw-gc 8 came along.  Repair
that.

Also while kite-language.org is still alive (and serving the source
tarball) the page itself just returns a mysql error.  archive.org
indicates that it's been like that for a year, so point the homepage
at an working version of it.

Given the state of the homepage and the fact that there hasn't been
a release since this bdw-gc patch was attempted to be upstreamed 9
years ago, I'm going to go ahead and mark this formula deprecated.
It looks like there has been no activity on this project since 2012.