imagesnap: remove explicit SDK setting from xcodebuild
This "-sdk" flag no longer seems to work; this meant that this
formula no longer builds from source and didn't get a Big Sur
rebottle.

The flag was originally added to this formula in 2011 to work
around some issue with Xcode 4 (3c8f29684e8b0842792d8312090f8bdcf5855a7c)
It seems to be the only formula that does this so it seem safe
to remove.

Also mark it as explicit public domain per:
  https://github.com/rharder/imagesnap/blob/master/README.md#a-note-about-public-domain